requirement.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
flake8==7.1.1
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.5
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
